pyproject.toml
[project]
name = "ml-project-template"
version = "0.1.0"
description = "ML Project Template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
    "google-cloud-bigquery>=3.11.0",
    "google-cloud-firestore>=2.11.0",
    "requests>=2.31.0",
    "google-generativeai>=0.3.0",
    "google-cloud-storage>=2.14.0"
]   
